SnapdragonCamera: Pictures were rotated when the device is in flat position
Normally, the picture taken by front camera needs to rotate 270 degree and
the picture taken by rear camera needs to rotate 90 degree. Launch camera
app when the device is in a close to flat position, camera app can’t get
device’s orientation. The picture will not be rotated when the orientation
is unknown.

Change-Id: I41d19f162f28aa6838b0246f14475480c8703e06
CRs-Fixed: 1063452